requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
boto3==1.24.59
botocore==1.27.59
Brotli==1.0.9
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudmersive-convert-api-client==3.2.5
cryptography==36.0.2
cycler==0.10.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.3
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-js-reverse==0.9.1
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.12.2
docx2pdf==0.1.7
docxtpl==0.11.3
drf-pdf==0.2.0
et-xmlfile==1.0.1
fonttools==4.34.4
h11==0.13.0
idna==3.3
inflate64==0.3.0
jdcal==1.4.1
Jinja2==2.11.3
jmespath==1.0.0
kiwisolver==1.3.1
kombu==5.2.4
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.5.2
multivolumefile==0.2.3
numpy==1.24.1
openpyxl==3.0.6
packaging==21.3
pandas==1.5.2
Pillow==9.4.0
prompt-toolkit==3.0.31
psutil==5.9.1
py7zr==0.20.0
pybcj==1.0.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.15.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyppmd==0.18.3
python-crontab==2.6.0
python-dateutil==2.8.1
python-decouple==3.6
python-docx==0.8.11
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2022.2.1
pyzstd==0.15.3
redis==4.3.4
reportlab==3.6.9
requests==2.28.0
s3transfer==0.6.0
sendgrid==6.9.7
shortuuid==1.0.9
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==2.2.0
texttable==1.6.4
timeago==1.0.16
tomli==2.0.1
tqdm==4.58.0
tzdata==2022.2
Unidecode==1.3.4
urllib3==1.26.9
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.3.0
wrapt==1.14.1
xlrd==2.0.1
